+++ examples/resources/seqera_google_credential/import-by-string-id.tf
@@ -1,7 +1,4 @@
-# Note: Import only works for user-level credentials
-# For workspace-level credentials, you must also set workspace_id in the config
-
 import {
   to = seqera_google_credential.my_seqera_google_credential
-  id = "1a2b3c4d5e6f7g8h9i0j"
+  id = "..."
 }